Add the ability to specify a timeout
If there is a non-zero timeout, we add timeout handling, so that
commands which _may_ take a _long_ time to execute don't hold up
the client.

Note, this **does not** affect the backend execution, the process
will continue to execute until completion or until you kill the
server.

mux.NewRouter returns a *mux.Router, which implements http.Handler.
To be able to conditionally substitute in a new Handler (if the
user specified a timeout - TimeoutHandler returns a Handler), I
changed the code to first create an interface variable of type
Handler, which I type inference to a *mux.Router as needed to add
the routes. This isrequired since http.Handler does not recognize the
HandleFunc, and since the concrete type is a pointer, I can be sure
that I'm accessing the same router each time.

If I were to just do router := mux.NewRouter(), for example,
router would be of type *mux.Router, so I wouldn't be able to just
swap out the return value of TimeoutHandler. I wouldn't even be able
to type inference the return of TimeoutHandler to a *mux.Router,
because the concrete type is **not** *mux.Router, it's *http.timeoutHandler!