Use go-bindata instead of storing the HTML directly in Go
Using go-bindata, we can store the HTML file separately in the
code, and then turn it into a Go file (to be put in the statically-
linked binary) to be referenced in the code just like any other
Go code.

This will make editing the HTML significantly easier since it's now
in its own file, not to mention that editors should now be able to
apply syntax highlighting to it.

The go.mod/sum additions were generated by the initial "go get",
and since its not directly used by this module, it's tagged as an
indirect dependency. Having a specific version in go.mod ensures
that we'll use the same version of go-bindata to generate our
bindata files every time.