Put command output in diff on current page instead
Prior to this commit: when you click a command, it takes you away
from the command station page and instead loads the data separately
in your browser, requiring that you click Back to kick off a different
command.

This commit adds a new diff to the page, and uses JS to "fetch"
the command output and modify the DOM to inject the text into the
div.